Move `get_user_agent` to `phoenix-channel`
