+++ rust/headless-client/src/main.rs
@@ -4,7 +4,6 @@ use anyhow::{anyhow, Context as _, Result};
 use backoff::ExponentialBackoffBuilder;
 use clap::Parser;
 use connlib_client_shared::{keypair, ConnectArgs, LoginUrl, Session};
-use connlib_model::get_user_agent;
 use firezone_bin_shared::{
     new_dns_notifier, new_network_notifier,
     platform::{tcp_socket_factory, udp_socket_factory},
@@ -15,6 +14,7 @@ use firezone_headless_client::{
 };
 use firezone_telemetry::Telemetry;
 use futures::{FutureExt as _, StreamExt as _};
+use phoenix_channel::get_user_agent;
 use phoenix_channel::PhoenixChannel;
 use secrecy::{Secret, SecretString};
 use std::{
